requirements.txt
tk
Pillow
beautifulsoup4
requests